+++ R/argument_validation.R
@@ -281,15 +281,14 @@ test_transition_probability_matrix <- checkmate::makeTestFunction(
 #' @export
 
 check_probability_vector <- function(
-    x, len = NULL, tolerance = sqrt(.Machine$double.eps)) {
+    x, len = NULL, tolerance = sqrt(.Machine$double.eps)
+) {
   checkmate::assert_number(tolerance, lower = 0)
-  res1 <- checkmate::check_atomic_vector(x, any.missing = FALSE, len = len)
-  if (!isTRUE(res1)) {
-    return(res1)
-  }
-  res2 <- checkmate::check_numeric(x, lower = 0, upper = 1)
-  if (!isTRUE(res2)) {
-    return(res2)
+  res <- check_numeric_vector(
+    x, any.missing = FALSE, len = len, lower = 0, upper = 1
+  )
+  if (!isTRUE(res)) {
+    return(res)
   }
   if (abs(sum(x) - 1) > tolerance) {
     return("Must add up to 1")
@@ -360,4 +359,63 @@ assert_list_of_lists <- checkmate::makeAssertionFunction(
 
 test_list_of_lists <- checkmate::makeTestFunction(
   check_list_of_lists
+)
+
+#' Check if an argument is a numeric vector
+#'
+#' @description
+#' This function checks whether the input is a numeric vector.
+#'
+#' @param x
+#' Object to check.
+#'
+#' @inheritParams checkmate::check_numeric
+#' @inheritParams checkmate::check_atomic_vector
+#'
+#' @return
+#' Compare to \code{\link[checkmate]{check_numeric}}.
+#'
+#' @export
+
+check_numeric_vector <- function(
+    x, lower = -Inf, upper = Inf, finite = FALSE, any.missing = TRUE,
+    all.missing = TRUE, len = NULL, min.len = NULL, max.len = NULL,
+    unique = FALSE, sorted = FALSE, names = NULL, typed.missing = FALSE,
+    null.ok = FALSE
+) {
+  res1 <- checkmate::check_atomic_vector(
+    x, any.missing = any.missing, all.missing = all.missing, len = len,
+    min.len = min.len, max.len = max.len, unique = unique, names = names
+  )
+  if (!isTRUE(res1)) {
+    return(res1)
+  }
+  res2 <- checkmate::check_numeric(
+    x, lower = lower, upper = upper, finite = finite, any.missing = any.missing,
+    all.missing = all.missing, len = len, min.len = min.len, max.len = max.len,
+    unique = unique, sorted = sorted, names = names,
+    typed.missing = typed.missing, null.ok = null.ok
+  )
+  if (!isTRUE(res2)) {
+    return(res2)
+  }
+  return(TRUE)
+}
+
+#' @rdname check_numeric_vector
+#' @inheritParams checkmate::assert_numeric
+#' @inheritParams checkmate::assert_atomic_vector
+#' @export
+
+assert_numeric_vector <- checkmate::makeAssertionFunction(
+  check_numeric_vector
+)
+
+#' @rdname check_numeric_vector
+#' @inheritParams checkmate::assert_numeric
+#' @inheritParams checkmate::assert_atomic_vector
+#' @export
+
+test_numeric_vector <- checkmate::makeTestFunction(
+  check_numeric_vector
 )