+++ tests/testthat/test-list_helpers.R
@@ -1,18 +1,26 @@
 test_that("lists can be merged", {
   expect_equal(
-    merge_lists(list("a" = 1, "b" = 2)),
-    list(a = 1, b = 2)
+    merge_lists(list("a" = 1, "b" = 2, "z" = NULL)),
+    list(a = 1, b = 2, z = NULL)
   )
   expect_equal(
-    merge_lists(list("a" = 1, "b" = 2), list()),
-    list(a = 1, b = 2)
+    merge_lists(list("a" = 1, "b" = 2, "z" = NULL), list()),
+    list(a = 1, b = 2, z = NULL)
   )
   expect_equal(
-    merge_lists(list("a" = 1, "b" = 2), list("b" = 3, "c" = 4)),
-    list(a = 1, b = 2, c = 4)
+    merge_lists(list("a" = 1, "b" = 2, "z" = NULL), list("b" = 3, "c" = 4)),
+    list(a = 1, b = 2, z = NULL, c = 4)
+  )
+  expect_equal(
+    merge_lists(list("a" = 1, "b" = 2), list("b" = 3, "c" = 4, "z" = NULL)),
+    list(a = 1, b = 2, c = 4, z = NULL)
   )
   expect_equal(
     merge_lists(list("a" = 1, "b" = 2), list("b" = 3, "c" = 4), list("d" = 5)),
     list(a = 1, b = 2, c = 4, d = 5)
+  )
+  expect_equal(
+    merge_lists(list("a" = 1, "b" = 2), list("b" = 3, "c" = 4), list("z" = NULL)),
+    list(a = 1, b = 2, c = 4, z = NULL)
   )
 })